requirements.txt
activity
aiofiles
aeval
Appium-Python-Client
async-timeout
asyncio
asyncpg
attrs
beautifulsoup4
bs4
certifi
cffi
config
Cybernator
dhooks
discord.py
discord-py-slash-command
discord-interactions
dnspython
Flata
form
ffmpeg
geojson
googletrans
Image
idna
idna-ssl
multidict
pafy
psutil
pillow
pycparser
pymongo
PyNaCl
pyowm
pytz
requests
Sequence
selenium
setuptools
six
soupsieve
tabulate
urllib3
utils
uuid
yarl
zope.interface
